+++ packages/awino-ui/src/components/buttons/ConnectButton.tsx
@@ -14,7 +14,7 @@ export default function ConnectButton(props) {
   }));
 
   const handleConnectToggle = useCallback(() => {
-    console.log('connecting/disconnecting');
+    // console.log('connecting/disconnecting');
   }, []);
   return (
     <Button color="primary" onClick={handleConnectToggle} {...props}>